+++ go/enclave/core/rollup.go
@@ -69,6 +69,8 @@ func EmptyRollup(agg gethcommon.Address, parent *common.Header, blkHash gethcomm
 		Number:      big.NewInt(int64(parent.Number.Uint64() + 1)),
 		// TODO - Consider how this time should align with the time of the L1 block used as proof.
 		Time: uint64(time.Now().Unix()),
+		// generate a random number inside the enclave that will be exposed to the evm
+		MixDigest: gethcommon.BytesToHash(RandomBytes(gethcommon.HashLength)),
 	}
 	r := Rollup{
 		Header: &h,