+++ go/enclave/evm/utils.go
@@ -25,9 +25,9 @@ func convertToEthHeader(h *common.Header) *types.Header {
 		Number:      h.Number,
 		GasLimit:    1_000_000_000,
 		GasUsed:     0,
-		Time:        uint64(0), // todo - pass in the block used as proof to get the timestamp
+		Time:        h.Time,
 		Extra:       obscuroHeader,
-		MixDigest:   gethcommon.Hash{},
+		MixDigest:   h.MixDigest,
 		Nonce:       types.BlockNonce{},
 		BaseFee:     gethcommon.Big0,
 	}