+++ docs/testnet/changelog.md
@@ -5,3 +5,5 @@
 * ObscuroScan transaction explorer preview launched.
 * Number Guessing Game smart contract deployed to Testnet.
 * Obscuro Docsite made available.
+
+* ``block.difficulty`` will return a true random number generated inside the secure enclave.